+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 ## 0.2.2 - 2023-01-02
-- worked send-report
+- defined send-report
 
 ## 0.1.1 - 2023-01-01
 - branch feature-ishige